Add Phases 9.5-9.8, ModelProbe example, app icon, and Phase 10 groundwork
- Phase 9.5 R1 clip pipeline / R4 hover; Phase 9.6 animation system
  (AnimationManager, CollapsiblePanel, ProgressBar, easing)
- Phase 9.7 adaptive layout (stretch/spacing/fillCrossAxis);
  Phase 9.8 AutoSize (GetPreferredSize/AutoSize, ResolveMeasurer seam,
  size-intent semantics)
- ModelProbe demo extracted to examples/ModelProbe/ (self-contained:
  main, sources, app.ico, ModelProbe.rc/.rc.in, own CMake target)
- CMake split into ECDI static library + modelprobe executable
- App icon (black rounded rect + white API) embedded via RC resources;
  window icon loaded in Win32WindowClass with graceful fallback
- Design docs for phases 9.5-9.8 and Phase 10 library-ization
  requirements (0.1.0: public API boundary, install/export,
  self-containment test, external consumer validation)
- gitignore: probe.exe exception (RCDATA build dependency),
  Python venv/build products excluded

+++ ECDI/src/Tests/TestFramework.cpp
@@ -151,8 +151,8 @@ void PrintSummary(const std::vector<TestResult>& results)
         }
     }
 
-    const int passed = std::count_if(results.begin(), results.end(),
-        [](const TestResult& r) { return r.passed; });
+    const int passed = static_cast<int>(std::count_if(results.begin(), results.end(),
+        [](const TestResult& r) { return r.passed; }));
     const int failed = static_cast<int>(results.size()) - passed;
     Logger::Log(LogLevel::Info, ToWide(std::string("----------------------------------\nTests: ") +
         std::to_string(results.size()) + "  Passed: " + std::to_string(passed) +
